src/utils/solidity_parser.py

import solcx
from typing import Dict, List, Any

class SolidityParser:
    def __init__(self):
        # Install specific solc version
        solcx.install_solc('0.8.0')
        self.solc_version = '0.8.0'
    
    def parse_file(self, file_path: str) -> Dict:
        """Parse Solidity file and return AST"""
        with open(file_path, 'r') as f:
            source = f.read()
            
        return self.parse_source(source)
    
    def parse_source(self, source: str) -> Dict:
        """Parse Solidity source code and return AST"""
        compiled = solcx.compile_source(
            source,
            output_values=['ast'],
            solc_version=self.solc_version
        )
        return compiled
    
    def extract_arithmetic_operations(self, ast: Dict) -> List[Dict]:
        """Extract arithmetic operations from AST"""
        operations = []
        
        def visit_node(node):
            if node.get('nodeType') in ['BinaryOperation']:
                if node.get('operator') in ['+', '-', '*', '/']:
                    operations.append(node)
                    
            for child in node.get('nodes', []):
                visit_node(child)
        
        visit_node(ast)
        return operations
    
    def extract_variables(self, node: Dict) -> List[str]:
        """Extract variable names from AST node"""
        variables = []
        
        def visit_node(node):
            if node.get('nodeType') == 'Identifier':
                variables.append(node.get('name'))
                
            for child in node.get('nodes', []):
                visit_node(child)
        
        visit_node(node)
        return variables
    
    def extract_external_calls(self, ast: Dict) -> List[Dict]:
        """Extract external calls from AST"""
        calls = []
        
        def visit_node(node):
            if node.get('nodeType') == 'FunctionCall':
                if self._is_external_call(node):
                    calls.append(node)
                    
            for child in node.get('nodes', []):
                visit_node(child)
        
        visit_node(ast)
        return calls
    
    def extract_state_changes(self, ast: Dict) -> List[Dict]:
        """Extract state-changing operations from AST"""
        changes = []
        
        def visit_node(node):
            if node.get('nodeType') == 'Assignment':
                if self._is_state_variable(node.get('leftHandSide')):
                    changes.append(node)
                    
            for child in node.get('nodes', []):
                visit_node(child)
        
        visit_node(ast)
        return changes
    
    def extract_functions(self, ast: Dict) -> List[Dict]:
        """Extract function definitions from AST"""
        functions = []
        
        def visit_node(node):
            if node.get('nodeType') == 'FunctionDefinition':
                functions.append(node)
                
            for child in node.get('nodes', []):
                visit_node(child)
        
        visit_node(ast)
        return functions
    
    def _is_external_call(self, node: Dict) -> bool:
        """Check if node represents an external call"""
        # Implementation details for external call detection
        return False
    
    def _is_state_variable(self, node: Dict) -> bool:
        """Check if node represents a state variable"""
        # Implementation details for state variable detection
        return False